requirements.txt
pandas==2.1.0
openpyxl==3.1.2
mysql-connector-python==8.1.0